backend/src/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-renderer-xlsx==0.4.1
et-xmlfile==1.1.0
Markdown==3.3.4
openpyxl==3.0.7
Pillow==8.3.1
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
XlsxWriter==3.0.1
